refactored core and core/utils imports
